Setup JWT and use passport jwt for signing in user and deleting post

+++ controllers/api/v1/posts_api.js
@@ -12,7 +12,7 @@ module.exports.index = async function(req, res){
                 path: 'user'
             }});
 
-    return res.json(200, {
+    return res.status(200).json({
         message: "List of posts",
         posts: posts
     })
@@ -21,17 +21,27 @@ module.exports.index = async function(req, res){
 module.exports.delete = async function(req, res)
 {
     try{
-        let posts = await Post.findByIdAndDelete(req.params.id)
+        let post = await Post.findById(req.params.id);
+        console.log(req);
+        console.log(post.user);
+        console.log(req.user.id);
+        if(post.user == req.user.id){
+            post.remove();
+            await Comment.deleteMany({post: req.params.id})
 
-    let comments = await Comment.deleteMany({post: req.params.id})
-
-    return res.status(200).json({
-        data:{
-            data: req.params.id
-        },
-        message: 'Post and associated comments deleted'
-        })
+            return res.status(200).json({
+                data:{
+                    data: req.params.id
+                },
+                message: 'Post and associated comments deleted'
+            })
+        }else{
+            return res.status(401).json({
+                message: "You cannot delete this post"
+            })
+        }
     }catch(err){
+        console.log('Error', err);
         return res.json(500, {message: "Internal Server Error"});
     }
